header comment added, int8_max

+++ code/dynamic_programming/rod_cutting/rod_cutting.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 #include <vector>
-
+// Part of Cosmos by OpenGenus Foundation
 using namespace std;
 int rodCutting(int rod_length, vector<int> prices){
     if (rod_length <= 0){
         return 0;
     }
-    int maximum_price = INT_MIN;
+    int maximum_price = INT8_MIN;
     for (int i = 0; i < rod_length; i++){
         maximum_price = max(maximum_price, prices[i] + rodCutting(rod_length-i-1, prices));
     }
